pyproject.toml
[tool.poetry]
name = "nautobot-data-validation-engine"
version = "1.0.0-dev"
description = "Provides UI to build custom data validation rules for data in Nautobot"
authors = ["Network to Code, LLC <opensource@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-data-validation-engine"
repository = "https://github.com/nautobot/nautobot-plugin-data-validation-engine"
keywords = ["nautobot", "data", "validation", "django"]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_data_validation_engine" },
]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
invoke = "*"
black = "*"
yamllint = "*"
bandit = "*"
pylint = "*"
pylint-django = "*"
pydocstyle = "*"
flake8 = "*"
nautobot = "^1.0.1"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

# Don't cache data for later comparisons
persistent="no"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"
# Allow "ot" (OnboardingTask instance) as an acceptable variable name
good-names="ot,"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
# Disabled due to noise: too-few-public-methods, too-many-ancestors, too-many-instance-attributes
disable = """,
    line-too-long,
    bad-continuation,
    too-few-public-methods,
    too-many-ancestors,
    too-many-instance-attributes,
    duplicate-code,
    raise-missing-from,
    super-with-arguments
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

